Initial non-Debian support
Adds initial, tentative support for Pacman and APK package managers.

Adds a norecursion feature to prevent Babashka from grabbing all .sh files in a tree, which allows per-OS-based selection to work.

Adds a default ignore of `files/`, to allow files that will be referenced by the configuration to not be loaded by Babashka.